refactor(utm): simplify UTM attribution handling by consolidating methods and updating service structure
